# docs/server/swagger.json
{
    "components": {
        "schemas": {
...
            "pkg_api_v1.UpdateRegistryRequest": {
                "description": "Request containing registry configuration updates",
                "properties": {
                    "allow_private_ip": {
                        "description": "Allow private IP addresses for registry URL or API URL",
                        "type": "boolean"
                    },
                    "api_url": {
                        "description": "MCP Registry API URL",
                        "type": "string"
                    },
                    "local_path": {
                        "description": "Local registry file path",
                        "type": "string"
                    },
                    "url": {
                        "description": "Registry URL (for remote registries)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "pkg_api_v1.UpdateRegistryResponse": {
                "description": "Response containing update result",
                "properties": {
                    "type": {
                        "description": "Registry type after update",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.getRegistryResponse": {
                "description": "Response containing registry details",
                "properties": {
                    "last_updated": {
                        "description": "Last updated timestamp",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of the registry",
                        "type": "string"
                    },
                    "registry": {
                        "$ref": "#/components/schemas/github_com_stacklok_toolhive-core_registry_types.Registry"
                    },
                    "server_count": {
                        "description": "Number of servers in the registry",
                        "type": "integer"
                    },
                    "source": {
                        "description": "Source of the registry (URL, file path, or empty string for built-in)",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/pkg_api_v1.RegistryType"
                    },
                    "version": {
                        "description": "Version of the registry schema",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.getServerResponse": {
                "description": "Response containing server details",
                "properties": {
                    "is_remote": {
                        "description": "Indicates if this is a remote server",
                        "type": "boolean"
                    },
                    "remote_server": {
                        "$ref": "#/components/schemas/registry.RemoteServerMetadata"
                    },
                    "server": {
                        "$ref": "#/components/schemas/registry.ImageMetadata"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.listServersResponse": {
                "description": "Response containing a list of servers",
                "properties": {
                    "remote_servers": {
                        "description": "List of remote servers in the registry (if any)",
                        "items": {
                            "$ref": "#/components/schemas/registry.RemoteServerMetadata"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "servers": {
                        "description": "List of container servers in the registry",
                        "items": {
                            "$ref": "#/components/schemas/registry.ImageMetadata"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.registryListResponse": {
                "description": "Response containing a list of registries",
                "properties": {
                    "registries": {
                        "description": "List of registries",
                        "items": {
                            "$ref": "#/components/schemas/pkg_api_v1.registryInfo"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "This is the ToolHive API server.",
        "title": "ToolHive API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/v1beta/registry": {
            "get": {
                "description": "Get a list of the current registries",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.registryListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List registries",
                "tags": [
                    "registry"
                ]
            },
            "post": {
                "description": "Add a new registry",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "501": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Implemented"
                    }
                },
                "summary": "Add a registry",
                "tags": [
                    "registry"
                ]
            }
        },
        "/api/v1beta/registry/{name}": {
            "delete": {
                "description": "Remove a specific registry",
                "parameters": [
                    {
                        "description": "Registry name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "No Content"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Remove a registry",
                "tags": [
                    "registry"
                ]
            },
            "get": {
                "description": "Get details of a specific registry",
                "parameters": [
                    {
                        "description": "Registry name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.getRegistryResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get a registry",
                "tags": [
                    "registry"
                ]
            },
            "put": {
                "description": "Update registry URL or local path for the default registry",
                "parameters": [
                    {
                        "description": "Registry name (must be 'default')",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/pkg_api_v1.UpdateRegistryRequest",
                                        "summary": "body",
                                        "description": "Registry configuration"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Registry configuration",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.UpdateRegistryResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Gateway - Registry validation failed"
                    },
                    "504": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Gateway Timeout - Registry unreachable"
                    }
                },
                "summary": "Update registry configuration",
                "tags": [
                    "registry"
                ]
            }
        },
        "/api/v1beta/registry/{name}/servers": {
            "get": {
                "description": "Get a list of servers in a specific registry",
                "parameters": [
                    {
                        "description": "Registry name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.listServersResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "List servers in a registry",
                "tags": [
                    "registry"
                ]
            }
        },
        "/api/v1beta/registry/{name}/servers/{serverName}": {
            "get": {
                "description": "Get details of a specific server in a registry",
                "parameters": [
                    {
                        "description": "Registry name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ImageMetadata name",
                        "in": "path",
                        "name": "serverName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.getServerResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get a server from a registry",
                "tags": [
                    "registry"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}